test elia

+++ main/java/playground/logic/Location.java
@@ -12,8 +12,8 @@ public class Location {
 	public Location(String location) {	//format "x,y"
 		super();
 		String[] s = location.split(",");
-		this.x = new Integer(s[0]);
-		this.y = new Integer(s[2]);
+		this.x = new Double(s[0]);
+		this.y = new Double(s[2]);
 	}
 	public Location(double x, double y) {
 		super();
